Revert "Fix hal_implementation_test to handle non-vintf stability"
This reverts commit 9039b24850f4ac4b598b7150dcd1440edb0f5b70.

The reason is that the approach for detecting the presence of the service
has been changed to use of a system property. Thus the interface
must not be stable and frozen because it is not vintf-stable.

Bug: 321651892
Bug: 323106963
Bug: 323989070
Test: atest hal_implementation_test
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:b7c327978d24ae51616fd677349c2a1e4f358ba3)
Merged-In: If3234ecc65025d86a2b3b32cfed9abe266885a26
Change-Id: If3234ecc65025d86a2b3b32cfed9abe266885a26