auto: Use audio effect config without preprocessing
The test `AudioEffectTest#ManageSourceDefaultEffects` fails on
GMS CF automotive because software hotword detector occupies
the primary input. This test used to be skipped on HIDL HAL
because it did not have preprocessing effects. This CL switches
the automotive to use the audio effects HAL configuration which
does not have preprocessing effects to match the HIDL configuration.

To be able to use a custom vendor config for the effects,
the effects configuration file is being removed from the audio
HAL VAPEX, thus it needs to be installed via 'device.mk' for
all CF targets.

Bug: 318423731
Test: run `atest audioeffect_tests` on cf_x86_64_auto-trunk_staging-userdebug
Change-Id: Ic0cf3a918f5b7f8a34dc55d7d43254ce8fac431b