Don't pass shared pointers to audio buffer around
Shared pointers can be copied and stored and the usual semantics of it
imply the underlying data is valid until the shared pointer is
destroyed. This is not true for audio buffers as those buffers are
released as soon as the function returns.

This changes those pointers into const references which explicity
indicate borrow semantics.

Bug: 327434508
Test: m webRTC
Change-Id: I9296390463002cbc0389b17777560ddfa9b98c9e