move ExtendedLauncherAction proto serialization into RPC function
Instead of eagerly serializing the proto and passing around a byte blob,
pass the proto around and serialize it at the last moment. This will
enable some follow up simplifications.

We now include a proto header in a header file. To make that work, I had
to add a "export_shared_lib_headers" to "libcuttlefish_command_util" and
change the linking of "libcuttlefish_command_util" from static to
shared.

Test: launch CF
Change-Id: I8c380ec0a900cc52058da69793b5656ae4343cb5