dissolve launcher_messager.h into command_util/util.h
These are two APIs for the same thing. When launcher_message is inlined
into the util.h functions, many things can be simplified.

The read/write code was tweaked a bit. There is less copying when
writing. When reading, the old code didn't check the bytes read, which,
surprisingly, is needed when calling the ReadExact{,Binary} functions.

Test: launch CF
Change-Id: Iddd0a01c2b491df40159d22983c749772430b7f4