Set multi-apex bootconfig property for keymint
This is done prior to actually setting up the multi-apex configuration,
for the benefit of the `v2/android-ota/cf_flashing_tot_to_but` test, in
the window where old host tools are run with a newer image.

Hopefully after this is submitted, it will be possible for the following
commit (aosp/3017011) to pass this test on presubmit.

Bug: 331474817S
Test: launch_cvd --noresume
Change-Id: I0b17bedce777a9543fdb77eb229c573544382969