fix `adb reboot` after snapshot restore
run_cvd was holding a `foreground_launcher_pipe` and assumed there was a
parent process holding the read half. However, when restoring, we
instead wait on a different pipe in the parent and then exit when the
restore appears to be complete. If the device later actually does see a
BOOT_COMPLETE event (e.g. because it rebooted), then it would try to
write to `foreground_launcher_pipe` and get a SIGPIPE.

This commit fixes that behavior by never creating the
`foreground_launcher_pipe` when restoring.

Bug: 332385095
Test: restore CF device, then `adb reboot`
Change-Id: I3013bbfa8b828df26f2846352c8f82bf5bc9c609