Add display 3 to CF display config
Display 3 was added to the cuttlefish emulator without specifying it in display_layout_configuration.xml. This leads to it being added to the default display/power group, making its power state tied to the main display, which is not desired. Specify display group for display 3 in the XML file to fix this.

Bug: 328289703
Test: adb shell settings put global android.car.DISPLAY_POWER_MODE 0:2,2:2,3:0 - only display 3 shuts off
Change-Id: I88d1c06a5459b2405588ceb95390242f40d9509d