Always attempt to download full set of system imgs
The current logic might only download the `system` and `product` images
from the system build's `img.zip`.  We want the full set of system
images if they exist.

Since we always download the target files if there is a `--system_build`
present we might as well skip the `img.zip` and extract them from the
target files.

Bug: 333616172
Test: cvd fetch --target_directory=/tmp/cvd/chd_target_files
--default_build=git_main/cf_arm64_only_phone-trunk_staging-userdebug
--system_build=git_main/cheetah-trunk_staging-userdebug
Test: cd /tmp/cvd/fetch_test
Test: bsdtar -xf system/<target_files_zip>
Test: shasum -b vbmeta_system.zip
Test: shasum -b IMAGES/vbmeta_system.zip
Test: # the hashes should match