Update key names to for clarity, add RSA2048 key
With future super image generating updates we will have a need for both
the RSA4096 and RSA2048 versions of the test keys.  Updated the RSA4096
key file name to make it clear which is which going forward.

While updating the paths, I consolidated them into a helper so any
future updates will only need to happen in a single location.

Bug: 322887496
Test: cvd_internal_start --daemon --resume=false
Change-Id: I87cbea1003782a325b2787c2a21681d4f50f34b8