Dynamically combine dynamic partitions and groups
Relanding after aosp/3060923 revert
Was using the wrong lookup key constant for a value, and added filtering
of the partition lists based on which images are actually extracted.
There was an issue with the vendor build having system images that did
not exist in the system build, so the `build_super_image` call was
failing.

Generate the dynamic partitions list and update groups from a
combination of both "dynamic_partitions_info.txt" files.

This is based on the `merge_target_files.py` logic, and generates a
close enough output for our purposes.

Bug: 322887496
Test: see the bug comment 37 for steps to test this stack of changes
Test: run v2/android-gsi/test_mapping_presubmit on CL to test

Change-Id: I343f79a056348f37012568800637aecaf913007a