fix adb proxy when powerwash is used on a restored device
`IsRestoring` was recently introduced to detect if the new device is for
a restore. If the device is reboot'd or powerwash'd, it correctly
returns `false`. However, the adb vsock proxy was still using its own
config option, `sock_vsock_proxy_wait_adbd_start`, that continues
returning `true` even after powerwash and so the adb proxy was
incorrectly waiting for a signal from some restore-only code. (this
isn't an issue for `adb reboot` because it doesn't restart the vsock
proxy process)

To fix, this commit deletes `sock_vsock_proxy_wait_adbd_start` and
instead uses `IsRestoring` directly.

Bug: 331741524
Test: m && (stop_cvd; launch_cvd --noresume --enable-sandbox=false --daemon --secure_hals=keymint,oemlock,guest_keymint_insecure && snapshot_util_cvd --subcmd=snapshot_take --auto_suspend --force --snapshot_path=$HOME/cf-snapshot && stop_cvd && launch_cvd --snapshot_path=$HOME/cf-snapshot) && powerwash_cvd
Change-Id: I493741948c2a3e87d41551820aa808c33d6cafc3