Adding fallback to previous system img.zip fetch
The fallback is intended to avoid a regression where the images are not
present in the target files but were in the img.zip.

It is also possible that we end up extracting the product.img twice with
this logic, but I image the case where the target files has the product
image but not the system image is rare enough that we are okay without
special handling for the case.