Make main crosvm wait for vhost user gpu device socket available
... to potentially avoid a race between the main crosvm connecting
on the socket that the gpu crosvm has not yet started accepting
connections on.

Crosvm's vhost user GPU seems to wait for a single connection and
then shuts down after that single connection exits. Because of this,
the existing WaitForUnixSocket() can not be used. schuffelen@
suggested the use of lsof which seems to be able to check the socket
state without actually connecting.

Also removes the creation of the socket as the vhost user gpu device
will create it and the launcher creating it ahead of time does not
really accomplish anything.

Bug: b/334188728
Test: ATP, see b/334188728#comment6
Change-Id: I7a05bf80d7afbe2658ad3af0c39876ef135a0d06