use `crosvm resume` to detect RESTORE_COMPLETE
Instead of waiting for logcat output, run the `crosvm resume` command to
block until the vCPUs have begun.

CvdBootStateMachine starts a thread to run that command. When the
command completes, it notifies the other CvdBootStateMachine thread via
a pipe, which then sends a value to the `--daemon` parent indicating
that the restore is complete (using the same mechanism as for a normal
boot complete).

Bug: 333111958
Test: m && (stop_cvd; launch_cvd --noresume --enable-sandbox=false --daemon) && snapshot_util_cvd --subcmd=snapshot_take --auto_suspend --force --snapshot_path=$HOME/cf-snapshot && stop_cvd && launch_cvd --snapshot_path=$HOME/cf-snapshot
Change-Id: Ia515c0485a7d6ec5cda611a7423674422b9f28d5