Attempt at addressing multiple identical fetches
Intended to avoid issues with multiple fetches interacting with the
upcoming local cache at the same time.  `SharedFD::Mkstemp` gives us a
unique filename without clashes, and `RenameFile` uses `rename` which
avoids problems with overwriting an existing file with the same name.

The drawback of this version is that anecdotally it seems to be 20-30%
slower during a basic `aosp-main` fetch.  I looked at using std::tmpnam
to create the filename and keep the `std::fstream` version, but that
function generates warnings because the filename creation and file
creation have a gap.

Need to look more into performance.

Test: cvd fetch --target_directory=/tmp/cvd/fetch_test
--default_build=aosp-main