Add deps to drmserver and mediaserver
We want to install different combinations of arch variants of the two
depending on the situations:

A: 64-bit only device: install 64-bit variants
B: 32-bit only device: install 32-bit variants
C: 64-bit/32-bit device (no GSI): install 32-bit variants
D: 64-bit/32-bit device (GSI): install both 32-bit and 64-bit variants

A, B, and C are achieved by adding the two modules to the new
multilib.prefer32.deps property.

D is achieved by adding the two modules to the multilib.both.deps
property with a select condition on
TARGET_DYNAMIC_64_32_DRMSERVER/MEDIASERVER properties.

Note that this is not complete. A few things to consider more.

* When the 32-bit variant is selected on a 64-bit supporting device
(i.e. the case C), the required deps of the 32-bit variant are not
installed, because the ABI mismatch.

* mediaserver and drmserver already have compile_multilib: "prefer32" in
their bp files. This disables the 64-bit variant even for case A and
causes a missing dependency error. We need to configure the target with
IGNORE_PREFER32_ON_DEVICE=true in order to keep the 64-bit variant.
Since we can't set the flag (globally) yet, add compile_multilib: "both"
so that the absence of 64-bit variant doesn't bother us.

Bug: N/A
Test: m aosp_cf_system_x86_64
Change-Id: I981a3397e4f06783957653b967f0ef50ef694e95