Move host artifact directory check to start cmd
And check for cvd_internal_start instead of cvd.

The check was previously made in the instance database. The start
command is the one that executes binaries from that directory and
therefore the one who should validate it. Moving the check out of the
database also allows to store devices before the fetch operation has
finished.