Enable Gfxstream's GlProgramBinaryLinkStatus after aosp/3151743
Prior to aosp/3151743, the host GLES translator was not tracking the
link status of programs created by glProgramBinary() and would always
return false for glGetProgramiv(... GL_LINK_STATUS ...) calls.

Also, prior to aosp/3151743, the guest GLES encoder was losing
information about `samplerExternalOES` between glGetProgramBinary()
and glProgramBinary() calls which would cause incorrect handling of
sampling from a binding with GL_TEXTURE_EXTERNAL_OES.

Guest applications seem to typically fallback to fully recreating
programs with shaders (glCreateShader() + glShaderSource() +
glAttachShader()) when linking fails with glProgramBinary(). This
lead to backwards compatibility problems when an old guest (which
does not have the above guest GLES encoder fix) runs with a newer
host (which does have the above host GLES translator fix) as the
fallback path would be disabled but the guest would have incorrect
GL_TEXTURE_EXTERNAL_OES handling.

With this, the corrected host behavior is hidden behind this feature
and the host feature should only be enabled with guests which also
have the guest fix. This change labels the current guest as having
the fix.

Bug: b/349821825
Test: GfxstreamEnd2EndTests
Test: cts -m CtsGraphicsTestCases
Test: cts -m CtsUiRenderingTestCases
Change-Id: I0d7543e4726cac09d571b3199139ee493ded17fc