Rename VmManagerBootconfig and remove outdated comment
aosp/1615297 created `VmManagerBootconfig` from
`VmManagerKernelCmdline` and copied the comment that came with it
which is a TODO comment to rearrange some `if` statements conditional on
the `vm_manager`.

`VmManagerBootconfig` has never had these conditionals and has a reduced
scope. Rename this function to match the true scope and remove the
comment.

Bug: b/352809765
Change-Id: I5c3633c0f3fd477cfb080e30127ab711b663263e
Test: m && cvd start --resume=false