Setup ethernet for cf_auto target.
This CL adds ethernet.rc which starts ethernet.sh to setup
ethernet during early boot for auto targets. The shell creates a
network namespace: auto_eth, moves the ethernet interface: eth0
into the netwwork namespace so that android network settings
will not interfere with it. The script assigns static IP and config
routing rule so that the guest could use it to communicate with
the VHAL server running on the host.

This CL adds required selinux rule.

This CL updated init.rc file for VHAL to include two VHAL services,
one is running vsock and one is running ethernet. The ethernet
service requires network namespace to be created, otherwise,
enter_namespace fails and the service will not be able to start.
We define a system property that is set after the network script
is run. We only enable the ethernet VHAL if we see the setup
is complete.

Ideally we only want to run vsock VHAL if we see the setup is not
completed, possibly due the network setup script is missing (in
GSI build where system_ext is replaced). However, init language
does not support checking whether a property is unset. Therefore,
we run vsock VHAL unconditionally and it will exit if it detects
the setup is completed.

Test: Local run on cf_auto, make sure ethernet is used.
dumpsys android.hardware.automotive.vehicle.IVehicle/default is working.

launch_cvd --num_instances 2: launch two instance with shared VHAL.
launch_cvd, In another terminal launch_cvd --instance_nums 2: launch
two instances with separate VHAL.

Bug: 353743121

Change-Id: If76ce9773f74fe1bd6520f06274346883a2eb082