16kb: Replace PRODUCT_BOOTS_16K for TARGET_BOOTS_16K
Use a target build variable to indicate that the target
boots with 16kb page sizes.

Target variables are available earlier in the build process, and they are needed to pick the right file system table. Product variables are defined after the selection of the file system table is done, that is the main reason we are replacing PRODUCT_BOOTS_16K.

Bug: 346868433
Test: Built Cuttlefish targets with 16kb support
      $ source build/envsetup.sh
      $ lunch aosp_cf_arm64_phone_pgagnostic-trunk_staging-userdebug
      $ get_build_var TARGET_BOOTS_16K
      true
      $ get_build_var PRODUCT_BOOTS_16K

Change-Id: If60740e180081ae32f6652d94cba8e106724711d