cvd fleet prints full info for non-running instances
By not relying only on the output of cvd_internal_status, but instead
keeping the relevant information in the database.
Some info is still parsed from the status command output the first time
it's run (right after a successful boot) and then stored in the
database: webrtc_device_id and adb_port.
Additionally, cvd fleet prints the adb port on its own, which is more
useful than the serial for remote use cases.