Open tap file descriptors in `crosvm` rather than `run_cvd`
This fixes networking in host-sandboxing mode.

When in a sandbox, `run_cvd` does not have permissions to negotiate with
`/dev/net/tun` to get file descriptors for tap devices. This change has
`run_cvd` give `crosvm` the name of the tap device, and `crosvm` sets up
the file descriptor itself. `crosvm` manages its own sandbox, so it is
not constrained the way `run_cvd` is.

The ioctls made by `run_cvd` and `crosvm` are comparable:

https://cs.android.com/android/platform/superproject/main/+/main:device/google/cuttlefish/common/libs/utils/network.cpp;l=109;drc=efb735f4d5a2f04550e33e8aa9485f906018fe4e

https://cs.android.com/android/platform/superproject/main/+/main:external/crosvm/net_util/src/sys/linux/tap.rs;l=174;drc=b06d296a1b34871683bd6684550441569d5da349

The error would be visible in this `run_cvd` output as well:
```
Unable to connect to cvd-wifiap-01 tap interface: Operation not permitted
Unable to connect to "cvd-wifiap-01": Bad file descriptor
Unable to connect to cvd-mtap-01 tap interface: Operation not permitted
Unable to connect to "cvd-mtap-01": Bad file descriptor
Unable to connect to cvd-etap-01 tap interface: Operation not permitted
Unable to connect to "cvd-etap-01": Bad file descriptor
```

Bug: b/318585008
Test: launch_cvd --resume=false, check networking
Test: launch_cvd --resume=false --enable_host_sandboxing, check networking
Change-Id: I612eddc8b15b4284925fb6c68fde532b3eef28bc