Re-enable printing on cuttlefish
Printing was originally disabled on cuttlefish in 2020 due to
unspecified failures.  Since then, the tests have been fixed.  Remove
printing from the list of disabled features so cuttlefish can be used
for printing development.  Since this was the last disabled feature in
cuttlefish_excluded_hardware, also remove the entire file and references
to it.

Bug: 356221362
Test: atest CtsPrintTestCases
Change-Id: Iea2b0de91a338f6bcd2b48a69e9d033060e406a0