Use ADB port provided by CVD in the operator
When cvd pregisters a device it includes the ADB port. This is more
reliable than depending on it being provided by the device at
registration time since that functionality was only introduced
recently and older branches don't have it. It still fallsback to the
device registration info to handle cases where the device wasn't started
by cvd.