Eliminate intermediary command
Cvd was using an unamed command in a file named handler_proxy.cpp to
basically separate cvd arguments from subcommand arguments. That command
would then create a new request and pass it to the actual command
handler.
This commit replaces that intermediary command with a call to a function
named ExtractCvdArgs that separates the arguments.