Reland: Re-enable printing on cuttlefish
Printing was originally disabled on cuttlefish in 2020 due to
unspecified failures.  Since then, the tests have been fixed.  Remove
printing from the list of disabled features so cuttlefish can be used
for printing development.  Since this was the last disabled feature in
cuttlefish_excluded_hardware, also remove the entire file and references
to it.

The timeouts that caused the previous revert should be resolved by ag/28852532 and aosp/3248612.

Bug: 356221362
Test: atest CtsPrintTestCases
Change-Id: Iffb208347ef007988a59acab50637828ad274f6e