Get rid of VerifyPrecondition
It was only checking that ANDROID_HOST_OUT was always present in the
environment, which is not always true. To make that check pass a default
path was calculated when the request was created and added to the
environment, only to be replaced with the value from an existing group
most of the time.