[skip ci] Duplicate some utilities from sandbox2
process_sandboxer was using these utilities even though the headers are
normally visibility-protected when the library is built with bazel.

https://github.com/google/sandboxed-api/blob/50d21e2627dd29d4ed18e163c8cf815e2f85cead/sandboxed_api/util/BUILD#L19

process_sandboxer could use these because the Android.bp port of the
build process does not respect the same visibility rules.

https://cs.android.com/android/platform/superproject/main/+/main:external/sandboxed-api/Android.bp;l=230;drc=3407ceb1183ca61c530a24170ed6a851b5506e4d

Using these utilities directly would get in the way when trying to build
this with bazel and a dependency on the sandbox2 library.

Bug: b/366513159
Test: m && process_sandboxer -- $(which cvd_internal_start) --resume=false --report_anonymous_usage_stats=y
Change-Id: I7156cb0faed0aacf255d61744a0b2bda4c620049
GitOrigin-RevId: e7dea13a39abea9dccde6518e292dda5939312d6