Skip multi-display system decoration tests for cuttlefish-md
It's necessary to skip multi-display system decoration tests for cars,
even if the device supports multi-display.
This is because the car only supports system bars on the user's main
display, but the tests use the user's secondary display, which doesn't
support system bars in cars.
Note that config_supportsSystemDecorsOnSecondaryDisplays is disabled by
default for car products.

Bug: 356518965
Flag: EXEMPT bugfix

Test: atest MultiDisplaySystemDecorationTests

Change-Id: I878e63f526d4bf9650ccff83fda34d7b66bb3be3