Protect access to the display sinks with the send mutex
The display_sinks_ property is modified when adding and removing
displays, and read when sending display buffers to the webrtc clients.

The sending of buffers can happen when a new frame arrives (no race here
as it's the same wayland server thread IIUC) and, crucially, from a
background thread that ensures frames are sent to the clients often.

Bug: 371590413
Test: locally
Change-Id: I8a48b3dfd7490b0cd4a4c2e4c23de88df5c0000e