Move non-vcpu/critical tasks to workers cgroup
In order to ensure proper emulation of SoCs from Cuttlefish when
vcpu_config are used, move various tasks to its own cgroup with cpusets
that are mutually exclusive from vCPUs. This assumes that the user
already has created some CGroup hiearchy that has the desired
configurations.

This is required because much of the device specific threads are spawned
from vCPU themselves and thus inheriting much of the vCPU thread's
properties such as cpu affinity. This can result in the device worker
threads competing for CPU slices, causing incorrect behaviours in the
guest. The device typically spin up their threads as part of MMIO
accesses from the vCPU itself, which occurs in the vCPU thread's
context. A much larger refactor is required to address the issue in
crosvm, this serves as a stop gap until a long term solution is in
place.

Test: m && boot up and saw that the threads migrated
Bug: 377969880
Change-Id: I6976e359b226491a41bfaab1ad8a6659445778ec
Signed-off-by: David Dai <davidai@google.com>