Remove the `merge_group` event from our actions
This change fixes a broken job and removes a duplicate job run.

The anchor and alias syntax appears to have broken the `artifacts` job.
That syntax might not be supported, or it is not done in the correct way
for Github's parsing.

For the merge queue, since "entering" the merge queue involves a
temporary branch being created we are already getting the `presubmit`
jobs triggered by `push`.  So PRs in the queue are double running jobs.

Bug: 341333157