Define vendor_ramdisk version of fstab.cf.* modules
*_available properties are used to allow reverse dependencies of
the module to depend on the specific partition variant of the module,
and should not be used to create the make module of the specific
partition for installation purpose. This change fixes the fstab.cf.*
modules by explicitly writing out the module definitions with
`vendor_ramdisk` property instead of creating the make module via
`vendor_ramdisk_available`.

Test: CI
Bug: 381104942
Change-Id: Ieaad869582d58f6b6e91361194911ea8259a8098