Add vendor_capabilities_service to the list of missing HALs
As of now this HAL implements feature that is only supported for
protected VMs. Additionally, the reference implementation that we are
going to provide will only work on devices running pKVM hypervisor.
Since cuttlefish doesn't support pVMs and doesn't run pKVM, add HAL to
the list of the known missing HALs.

Bug: 360102915
Test: presubmit
Change-Id: I9401ff64d3b3892b0cdc02a6d8bbb1fc634a77fd