Handle errors from WalkDirectory
Update the call stack to propagate the error from WalkDirectory. When
called recursively, propagate the error at the first instance. From
inspection of the users of the code, none seem to rely on the previous
behavior (to be allowed to "skip" recursive directory walks without
error), so we might as well harden the code while fixing the warning.

Bug: 385004915
Test: m
Change-Id: I9aa287f6ff0125ff2dffa09d1cdd43df9aeda18b