Initial improvements to `GetDiskUsage`
The `-k` flag was shadowing part of the `-b` flag's functionality.
Removed the multiplication in the `return` statement and set the `du`
output to be in bytes rather than kilobytes.
Added comments to avoid manpage reading.

https://man7.org/linux/man-pages/man1/du.1.html

Converted to `CF_EXPECT` over `CHECK` and `ParseInt` over `atoi` to
match current team practices.