Fixes for clang-tidy 19 warnings
```
The 1st argument to 'linkat' is -1 but should be a valid file descriptor or AT_FDCWD [clang-analyzer-unix.StdCLibraryFunctions,-warnings-as-errors]
...
linkat(-1, name.c_str(), AT_FDCWD, path.c_str(), AT_SYMLINK_FOLLOW);
...
The 1st argument to 'getcwd' is NULL but should not be NULL [clang-analyzer-unix.StdCLibraryFunctions,-warnings-as-errors]
...
std::unique_ptr<char, void (*)(void*)> cwd(getcwd(nullptr, 0), &free);
```

I'm not convinced by the legitimacy of these particular warnings.
https://man7.org/linux/man-pages/man2/link.2.html writes "If oldpath is
absolute, then olddirfd is ignored." and clang-tidy doesn't appear to
acknowledge this case.

For `getcwd`, it appears we are relying on a glibc extension to POSIX
that clang-tidy is ignoring.
https://man7.org/linux/man-pages/man3/getcwd.3.html writes "As an
extension to the POSIX.1-2001 standard, glibc's getcwd() allocates the
buffer dynamically using malloc(3) if buf is NULL."

Fixed both of these cases anyway as I have an upcoming change to
github.com/google/android-cuttlefish that upgrades clang-tidy and LLVM
in general to version 19.1.0, which raises these warnings.

Bug: b/385219027
Change-Id: Ibbadede0a94c264167bc2cb42cdf6ec4317796a1
Test: `cvd create`