Add scoped logger to avoid manual logger managing
The manual managing leaves more opportunity for a mistake or error to
cause the previous logger to never be reset.

Also, for tee logging we always wanted to set the Android LogSeverity to
`VERBOSE` because the levels are managed internal to those loggers.
This is automatically handled as well, which is why the
`ScopedLogSeverity` is a member instead of being instantiated alongside
the logger.

The new `LogToStderr` (and `ScopedTeeLogger`) will be used in the
upcoming `cvd cache`, because it will eventually be called from `cvd
fetch` invocations.

Test: TBD, I think a `cvd load` call will have successive logger setups?