Define vintf_fragments as modules
Define vintf_fragments as modules if the file is shared among multiple
modules to avoid any conflicts.

Note that additional vintf_fragment module is defined for the purpose
of installation in recovery partition. Setting `stem` property is not
needed as vintf_fragment module installs file based on the source file
name, not the module name.

Test: unpack and diff soong and make generated recovery partitions
Bug: 322089980
Bug: 381888358
Change-Id: I64648d443dbea75c7d854b2efe51ab09f134d9a6