Partially import libcuttlefish_command_util from AOSP
This is a stripped down version of the original as some dependencies are
not yet in this repository. Only entire functions were removed from this
version of the library so the interface and implementation of the
imported bits are the same as the original; therefore when the full
version is imported the code that uses this version will continue
working correctly.