Update file helpers
Had an error with my usage of `FileModificationTime` that was silently
failing and using the default value.

The new `DirectoryContentsPaths` is needed to correct that error, where
I was calling `FileModificationTime` with only a filename and not a
filepath.  There are also usages of `DirectoryContents` with the
AOSP version of this library where this new helper would be better
suited.