CF: Create boilerplate test for HWC test using CF
[Why]
We want to perform more testing on HWC beyond what VTS can do. One way
of doing this is to leverage Cuttlefish(CF) capabilities. CF allows to
simulate hotplugs, take screenshots, and other tasks that come close to
what an end user would do.
context: go/auto-display-testing-android

[How]
This is not supposed to be a very meaningful test, but rather a simple
boilerplate for more feature intensive tests to follow.
1. Create a simple binary (simply_red.cpp) that display a solid red
   color on the device. This binary takes control over the HWC and
   communicates with it through its AIDL interface. The binary makes use
   of Composer Client Wrapper, utilized by HWC VTS.
2. Create a simple java host test that runs this binary, takes a
   screenshot of the device and verify the color.

Bug: b/383783838, b/383782707, b/383783834
Test: atest CuttlefishHwcTests
Flag: TEST_ONLY
Change-Id: Ic3767fa23710468c7d317902d52f81958c69ebc4