sepolicy: Allow access to a second graphics card
With the use of VKMS, we have two drivers that create devices files for
the graphics card, virtio_gpu.ko and vkms.ko. VKMS cannot fully replace
virtio_gpu, so they have to co-exist. That means two device files will
be created.
Change the SELinux policy so that is allowed.

Bug: 383087740
Bug: 392117008
Test: Build and successfully boot Cuttlefish.
      Make sure there are no build or kernel errors regarding vkms.
      Make sure there are selinux issues regarding vkms or /dev/dri/card1

Change-Id: I43b2b11235caa1dda7d7b487a317336658c2ed38
Signed-off-by: Paz Zcharya <pazz@google.com>