Create one location to set the C++ standard flag
Will make updating the flag value a small change going forward.  In the
future we should standarize the `copts` values across our builds so the
flags/warnings/errors all stay consistent, but this commit is only to
swap to using the variable and not to make changes.

Decided to not update `libbase` or `libsparse` since that code is a copy
and I do not believe the team has any intention of making updates to it.

Bug: 388011569