Fix the genrule for `cvd version` output
The build runs in a sandbox and not in the working directory the `bazel`
command is executed in, so the previous version never had an output to
populate the `cvd version` field.  Also, `git describe` was returning an
old tag.  I updated to return the latest HEAD commit hash instead.

If the
`--working_status_command=path/to/tools/buildutils/stamp_helper.sh` flag
is added to a build, the output goes into a file intended for embedding
outside information for use inside the build sandbox.

The commit hash should not change "often" (like a timestamp would be
different every single build), so we mark it as a stable key.  Then, if
the value changes, every affected action (marked `stamp = 1`) is
triggered.  Volatile keys do not trigger actions when they change.