Radio: bump radio AIDL services
Bump radio AIDL version, and implement check for bands and channels on
startNetworkScan for E-UTRAN, since a new check was added in
RadioNetworkTest to verify that the channels provided as argument
correspond to the provided bands. This test only became active when
bumping the AIDL interface versions.

Bug: 393444294
Test: lunch aosp_cf_x86_64_phone-trunk_staging-userdebug && m
Test: cvd create -modem_simulator_sim_type=2
Test: atest VtsHalRadioTargetTest && atest CtsCarrierApiTestCases
Flag: EXEMPT no logic change
Change-Id: I020d16dfd292896d4009da34a0e7a48afa0ef98d