Integrate CF sensor sub-HAL
Integrate CF sensor sub-HAL which reuses most part of Goldfish sensor
sub-HAL except for the transportation.

Introduce SensorsHalProxy to handle the communication with the HAL:
SensorsHalProxy runs 2 threads in loops to (1) respond to the guest
requests and (2) report sensor data to the guest at a constant rate.

For now, SensorsHalProxy only supports "list-sensors" request, which is
required to activate sensor HAL. The support for more HAL requests can
be added later.

Bug: 350076857
Test: check logcat to confirm the guest receives sensors data
Change-Id: I1cd8004a6cb1d885007fbf43fd9a1cb06d1a7b90