Re-activate sensor HAL once the device reboots
When the guest VM reboots, the host needs to stop sending sensor
reports to the guest until the sensor HAL completes the sensor
discovery process again and re-activates. Otherwise, the continious
sensor reports from the host would interfere with the sensor discovery
process, causing the sensor HAL activation to fail in a reboot.

Bug: 350076857
Test: confirm sensor HAL is re-activated after `adb reboot`
Change-Id: I02329608bc766dfa289bd1cc5139b5611e615546