Build libcuttlefish_metrics
And the new metrics headers target.  Introduced to allow this library
and the `metrics` binary to depend on them without introducing a new
dependency between `libcuttlefish_metrics` and `metrics`.

Does have the minor downside of needing to modify include paths
slightly.

Swapped `strcpy` to `strncpy` to fix a `clang-tidy` warning.

Bug: 402249059