Import and install extract-vmlinux
This change introduces a bit of a mess with installed files.
extract-vmlinux is renamed to extract-vmlinux.sh because bazel
requires that targets and source files have distinct names.

Additionally, bazel's sh_binary is implemented to create a symlink to
the source file. This is inconvenient for dh_install which copies over
the symlink into the bazel install directory without resolving it.
Therefore, this change adds a dh_install path directly into the source
tree to the now renamed extract-vmlinux.sh file, and sets up a symlink
with dh_link to create the original correctly named extract-vmlinux
file in the output directory.

Bug: b/402283877