Move HWC Tester to platform_testing
[Why]
HWC Tester is a useful util to be used in other tests beyond CF.
In platform_testing/, it will also run in conjuction with VKMS Tester.

[How]
Move hwc_tester.* to platform_testing/tests/display/hcct/utils.
With the util moved, remove the util directory and put the tests
directly in src/
Adjust the headers and Android.bp for the change.

Bug: b/398831713, b/352790467
Test: atest CuttlefishHwcRedSmokeTest
Flag: TEST_ONLY
Change-Id: I33906bd14c845cb48753aa710fc8a89eb50cc31a