Fix 100% CPU uage in webRTC process
The next_send variable is initialized to now() plus an offset and is
only subsequently set in the loop over display_last_buffers_. Because
the latter assignment uses std::min to compare against the existing
value of next_send the value never changes because the value it was
initialized to will always be earlier. As a result the wait_until()
call in the next iteration of the loop returns immediately and we
busy loop. Fix the issue by setting next_send to a reasonable maximum
before looping over display_last_buffers_.

Change-Id: I0adf236ddcdcc968ee3a0538154482d9bdfad3ad