Swap remaining non-orchestrator cases
One of the e2etests fails if they are changed.
The build from `aosp-android-latest-release` is missing the flag from aosp/3451951.

Since the swap is not pressing, we can stay on the frozen `aosp-main`
build.  Once the next quarterly release comes out, or we use the CF host
tools from the debian package in tests, we can swap the rest.