Fix environment for `cvd start --help` `Command`
Mostly this change is for the fix in `ConstructCvdHelpCommand` to modify
the help `Command`'s environment and not the (irrelevant in this case)
`cvd` invocation environment.

This matches the `ConstructCvdGenericNonHelpCommand` behavior and having a
relevent `ANDROID_HOST_OUT` will matter for following changes around
calling `cvd start --help`.

Since we are not modifying the `cvd` invocation `cvd_common::Envs`
value, `ShowLaunchCommand` would have to be updated.  The output seemed
useful enough to include anywhere we were printing, so I pushed it into
`Command::operator<<`.

Note: the lookup loop implementation is reversed because the environment
variables and their values are stored as strings in the form
"<var>=<value>" instead of in a map.